+++ HISTORY.md
@@ -16,7 +16,8 @@
     * block-instances can be dragged off from templates in the "export blocks", "unused blocks" and "hide blocks" dialogs
     * added "enter" key to key-pressed dropdown
     * added green flag symbol to message drop-down
-    * the green flag button's background color indicates whether the current scene is running or idle 
+    * the green flag button's background color indicates whether the current scene is running or idle
+    * empty categories are indicated by half-tone buttons
     * new preference setting per scene for pen color model and graphic effects, HSV or HSL, default is HSV
     * new preference setting per scene to "disable click-to-run" on blocks, for use in micro-world extensions
 * **Notable Changes:**
@@ -61,6 +62,9 @@
     * German
     * Chinese, thanks, Simon!
     * Brazilian Portuguese, thank you, Cassiano D'Andrea!
+
+### 2021-11-19
+* objects, byob, gui: visually indicate empty categories by half-toning their buttons
 
 ### 2021-11-18
 * gui: only pause generic hat blocks when loading a new project, not when switching to a new scene